packages/data/models/pyproject.toml
[project]
name = "models"
version = "0.0.1"
description = "OpenRouter Models"
readme = "README.md"
requires-python = "==3.14.2"
dependencies = [
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
]
